pyproject.toml
[tool.poetry]
name = "assistente-copel"
version = "0.1.0"
description = ""
authors = ["João H. L. Corrêa <jhenriquelc@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-generativeai = "^0.8.3"
google-ai-generativelanguage = "^0.6.10"
flask = "^3.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
